Tfoodie backend/add-recetas-favoritos/requirements.txt
# Function dependencies, for example:
# package>=version
pymongo>=3.10.1
dnspython>=1.16.0
